+++ HISTORY.rst
@@ -25,6 +25,7 @@ History
 - Reintroduce headers in settings page of new schematisation wizard
 - Run auto refresh on regaining window focus
 - Make breadcrumbs widget clearer with dropdown for long paths
+- Reorganize FileView widget with collapsible groups
 - Show revision number in RevisionView
 - Use `show_authentication` option in activating Rana menu to show authentification related settings
 - Scenario: zip file containing log files now properly extracted (#3302)